Add files via upload
Used to compare the two resultant columns from the GLDS Comparison Template - Unique Genes sheet. The input is the list of genes in the A and D columns into files called "input1" and "input2", respectively. The output files are a unique list of A vs D, D vs A, and the combined overlap. These outputs are then entered into the B, E, and H columns in the GLDS Comparison Template - Unique Genes sheet.